Add message parameters to API response
We previously added the message key as .message property
(97b433286dc086ea5261b1b88736600b5454887f), leaving it for the API
client to know what the parameters are and how to get them.
(Here, the parameters are the .line property and the page title.)

Let's just do this work for the client, so that it doesn't have to
know what TitleBlacklist is - it'll be able to just display the given
message with the given parameters. My specific use case is the upload
dialog in core (core shouldn't have to know about any extensions).

See also AbuseFilter change I5780eae96930211191ecd874aacf53fdacb58f89.

Change-Id: I97c1f5c6bbbdfc0b8ea9914bb075d5299c14df8f