Remove redundant documentation from test
We generally do not put these instructions in every test. A README file
would be more appropriate, but since there isn't one, I'm not sure if
it's needed, especially because the instructions for running tests
should ideally be the same for all extensions, and those could be
documented in a single place.

In particular, remove reference to soon-to-be-deprecated phpunit.php and
other information that can easily be found in the first 10 lines of
code.

Bug: T90875
Change-Id: I97e6f608f8a3f11febf33c38811fe7692be59e24