Replace confusing count() on associative arrays with known keys
Calling count() on an array with string keys technically works. But
it's confusing and ultimately meaningless. An array might contain
2 elements, but we have no idea how they are named. The code
continues to access some very specific elements, but really just
assumes they are there. We are lucky this never crashed.

We have very powerful language features like isset() or ?? to
express much more precisely what this code is supposed to do.

Change-Id: I7a37371b577c1e44a3c8514316198cf0f4288250